Move `nightly_options` module to its own file
